+++ client/lib/gin1.dart
@@ -1,12 +1,7 @@
-import 'package:client/gin3.dart';
 import 'package:flutter/material.dart';
-// import 'widgets/gin_widget.dart';
+import './gin3.dart';
 import './gin2.dart';
 import './service/login/google_login_service.dart';
-
-void main() {
-  runApp(const App());
-}
 
 class App extends StatelessWidget {
   const App({super.key});
@@ -84,19 +79,6 @@ class gin1 extends StatelessWidget {
                 ],
               ),
             ),
-
-            // ElevatedButton.icon(
-            //   onPressed: () {
-            //     Navigator.pushNamed(context, '/gin2');
-            //   },
-            //   icon: const Icon(Icons.account_circle),
-            //   label: const Text('Google 아이디로 로그인'),
-            //   style: ElevatedButton.styleFrom(
-            //     padding: const EdgeInsets.symmetric(
-            //         horizontal: 30.0, vertical: 30.0),
-            //     textStyle: const TextStyle(fontSize: 18),
-            //   ),
-            // ),
             const SizedBox(height: 10),
             ElevatedButton.icon(
               onPressed: () {
